Split book collection as a module
