+++ index.js
@@ -1,55 +1,6 @@
 /* eslint-disable max-classes-per-file */
-const inputTitle = document.getElementById('title');
-const inputAuthor = document.getElementById('author');
-const submitBtn = document.querySelector('.add-btn');
-const bookSection = document.querySelector('.books');
-class Collection {
-  constructor(books = []) {
-    this.books = books;
-  }
 
-  add(data) {
-    this.books.push(data);
-    this.display(data);
-    this.remove();
-    this.populateStorage();
-    inputAuthor.value = '';
-    inputTitle.value = '';
-  }
-
-  remove() {
-    const removeBtns = document.querySelectorAll('.remove-button');
-    removeBtns[removeBtns.length - 1].addEventListener('click', (evt) => {
-      this.removeFromColl(evt.target);
-      bookSection.removeChild(evt.target.parentNode);
-    });
-  }
-
-  display(data) {
-    if (this) {
-      const div = document.createElement('div');
-      div.className = 'book-wraper';
-      div.innerHTML = `<h3>"${data.title}" by ${data.author}</h3>
-                    <button data-value="${data.title}-${data.author}" type="button" class ="remove-button">Remove</button>`;
-      bookSection.appendChild(div);
-    }
-  }
-
-  removeFromColl(data) {
-    const arr = data.getAttribute('data-value').split('-');
-    this.books = this.books.filter(
-      (item) => item.title + item.author !== arr[0] + arr[1],
-    );
-    this.populateStorage();
-  }
-
-  populateStorage() {
-    localStorage.setItem(
-      'bookCollection',
-      JSON.stringify({ bookColl: this.books }),
-    );
-  }
-}
+import Collection from './modules/collection.js';
 
 class Book {
   constructor(title, author) {
@@ -57,6 +8,11 @@ class Book {
     this.author = author;
   }
 }
+
+const inputTitle = document.getElementById('title');
+const inputAuthor = document.getElementById('author');
+const submitBtn = document.querySelector('.add-btn');
+const dateBox = document.querySelector('.date');
 
 const coll = new Collection();
 if (localStorage.getItem('bookCollection')) {
@@ -69,56 +25,6 @@ if (localStorage.getItem('bookCollection')) {
 submitBtn.addEventListener('click', () => {
   coll.add(new Book(inputTitle.value, inputAuthor.value));
 });
-
-/* document.addEventListener('click', (e) => {
-  const { target } = e;
-  if (!target.matches('nav a')) {
-    return;
-  }
-
-  e.preventDefault();
-  urlRoute();
-});
-
-const urlRoutes = {
-  '/': {
-    template: '/index.html',
-    title: '',
-    description: '',
-  },
-  '/Add': {
-    template: '/add.html',
-    title: '',
-    description: '',
-  },
-  '/contact': {
-    template: '/contact.html',
-    title: '',
-    description: '',
-  },
-};
-
-const urlRoute = (event) => {
-  event = event || window.event;
-  event.preventDefault();
-  window.history.pushState({}, '', event.target.href);
-  urlLocationHandler();
-};
-
-const urlLocationHandler = async () => {
-  const location = window.location.pathname;
-  if (location.length == 0) {
-    location = '/';
-  }
-
-  const route = urlRoute(location);
-  const html = await fetch(route.template).then((response) => response.text());
-  document.getElementById('content').innerHTML = html;
-};
-
-window.onpopstate = urlLocationHandler;
-window.route = urlRoute;
-urlLocationHandler(); */
 
 const navAdd = document.querySelector('#add-new');
 const navList = document.querySelector('#list');
